pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "promptbox"
version = "0.2.0"
authors = [
    { name="AI Software Architect", email="developer@example.com" },
]
description = "A Streamlit application for managing, testing, and interacting with LLM prompts, characters, and scenarios."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "streamlit>=1.33.0",
    "sqlalchemy>=2.0.29",
    "requests>=2.31.0",
    "langchain>=0.1.16",
    "langchain-community>=0.0.34",
    "langchain-mistralai",
    "langchain-groq",
    "langchain-google-genai",
    "langchain-openai",
    "openai>=1.25.1",
    "mistralai>=0.1.6",
    "groq>=0.5.0",
    "google-generativeai>=0.5.2",
    "ollama>=0.1.8",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.1",
]
